Moved LIVD mappers to their own file
