+++ prime-router/src/main/kotlin/cli/LivdTableDownload.kt
@@ -7,8 +7,8 @@ import com.github.ajalt.clikt.parameters.options.flag
 import com.github.ajalt.clikt.parameters.options.option
 import com.github.ajalt.clikt.parameters.types.choice
 import com.github.ajalt.clikt.parameters.types.file
-import gov.cdc.prime.router.LivdTableColumns
 import gov.cdc.prime.router.common.Environment
+import gov.cdc.prime.router.metadata.LivdTableColumns
 import it.skrape.core.htmlDocument
 import it.skrape.fetcher.HttpFetcher
 import it.skrape.fetcher.response
